refactor: infra -> infrastructure 패키지명 변경
