+++ back/src/main/java/com/ll/goohaeyou/payment/payment/application/PaymentService.java
@@ -17,7 +17,7 @@ import com.ll.goohaeyou.payment.payment.domain.Payment;
 import com.ll.goohaeyou.payment.payment.domain.repository.PaymentRepository;
 import com.ll.goohaeyou.payment.payment.domain.type.PayStatus;
 import com.ll.goohaeyou.payment.payment.exception.PaymentException;
-import com.ll.goohaeyou.payment.payment.infra.TossPaymentUtil;
+import com.ll.goohaeyou.payment.payment.infrastructure.TossPaymentUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.minidev.json.JSONObject;
